Always use pull request date for version date
This commit fixes two issues:

1. If a git commit is amended and pushed after the PR is open, the put
to update the PR status will trigger additional builds. This is because
the date in the version in out is from the git commit which does not
match the PR date output by check.

2. In progress PRs were never skipped because in check containers
BUILD_ variables are not set. RESOURCE_ variables are set instead but
there is no guarantee that the resource name matches the job name. This
change makes check search comments using a smaller string that does not
depend on BUILD_ variables.